ourpursuit/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cycler==0.12.1
Django==4.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
httplib2==0.22.0
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.3
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pandas==2.2.1
Pillow==10.1.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.3.0
tqdm==4.66.2
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
validate-email-address==1
whitenoise==6.6.0
